+++ src/main/java/de/samply/exporter/ExporterConst.java
@@ -149,6 +149,8 @@ public class ExporterConst {
     public static final String ARCHIVED_INQUIRIES = "/archived-inquiries";
     public static final String ERROR_INQUIRIES = "/error-inquiries";
     public static final String LOGS = "/logs";
+    public static final String TEMPLATE_IDS = "/template-ids";
+    public static final String TEMPLATE = "/template";
 
     public static final String[] REST_PATHS_WITH_API_KEY = new String[]{CREATE_QUERY, RETRIEVE_QUERY,
             QUERIES, REQUEST, ACTIVE_INQUIRIES, ARCHIVED_INQUIRIES, ERROR_INQUIRIES, INQUIRY,