+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Divide Excel file in several files if number of rows is too large
 - CQL to Bundle Converter
 - Clear Log Buffer
+- Get Template and list of templates
 
 ### Changed
 - Move FHIR Packages to dktk-exporter